pyproject.toml
[tool.poetry]
name = "flm"
version = "0.3.0alpha000008" # ALSO BUMP IN flm/flm/__init__.py
description = "Flexible Latex-like Markup"
authors = ["Philippe Faist"]
license = "MIT"
include = ["flm/main/templates/**/*"]

[tool.poetry.scripts]
flm = 'flm.__main__:run_main'

[tool.poetry.dependencies]
python = "^3.8"
pylatexenc = {'git' = "https://github.com/phfaist/pylatexenc.git", branch = "devel" }
python-frontmatter = "^1.0.0"
PyYAML = ">=5.0"

[tool.poetry.dev-dependencies]
pytest = "^7.1.2"
ipython = "^8.3.0"
toml = "^0.10.2"

[tool.poetry.group.buildjslib]
optional = true

[tool.poetry.group.buildjslib.dependencies]
Transcrypt = ">=3.9.0"
PyYAML = ">=5.0"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"




[tool.pytest.ini_options]
testpaths = [
    "test",
]
